Updating neon and fixing a weird bug
