--- mpc-wallet/mpc-wallet-nodejs-win/.gitignore
@@ -1,2 +1,0 @@
-native/index.node
-native/artifacts.json
