--- mpc-wallet/mpc-wallet-nodejs-win/lib/index.js
@@ -1,2 +1,0 @@
-const MPCwallet = require('../native/index.node');
-console.log(MPCwallet.dh_init(10, "Secp256k1"));
