--- mpc-wallet/mpc-wallet-nodejs-win/native/build.rs
@@ -1,3 +1,0 @@
-fn main() {
-    neon_build::setup();
-}
