+++ mpc-wallet/mpc-wallet-nodejs/.gitignore
@@ -1,2 +1,5 @@
-native/index.node
-native/artifacts.json
+target
+index.node
+**/node_modules
+**/.DS_Store
+npm-debug.log*
