--- mpc-wallet/mpc-wallet-nodejs/native/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "mpc-wallet-nodejs"
-version = "1.1.3"
-build = "build.rs"
-edition = "2018"
-
-[lib]
-name = "mpc_wallet_nodejs"
-crate-type = ["cdylib"]
-
-[features]
-default = ["secp256k1"]
-k256 = ["nash-mpc/k256"]
-secp256k1 = ["nash-mpc/secp256k1"]
-
-[build-dependencies]
-neon-build = "0.4"
-
-[dependencies]
-# FIXME: need to pin neon to 0.4.0 until https://github.com/GabrielCastro/neon-serde/issues/61 is fixed
-neon = "=0.4.0"
-neon-serde = "0.4"
-nash-mpc = { version = "*", path = '../../nash-mpc', default-features = false, features = ["rust_gmp"] }
-serde_json = "1.0"
-
-[profile.release]
-# optimize for speed rather than for binary size
-opt-level = 3
