--- mpc-wallet/mpc-wallet-nodejs/native/build.rs
@@ -1,3 +1,0 @@
-fn main() {
-    neon_build::setup();
-}
