+++ mpc-wallet/mpc-wallet-nodejs/tests/index.js
@@ -1,0 +1,2 @@
+const MPCwallet = require('..');
+console.log(MPCwallet.dh_init(10, "Secp256k1"));
